+++ CHANGES.md
@@ -4,7 +4,8 @@
 
 - Separated file upload to its own component
 - File upload now supports drag and drop
-- Orpan file downloads are now streamed
+- Orphan file downloads are now streamed
+- Added ability to delete orphaned files
 
 ## 1.3.10